requirements.txt
certifi==2020.12.5
lxml==4.6.3
PDFNetPython3==8.1.0
Pillow==8.1.2
python-docx==0.8.10
reportlab==3.5.66
wincertstore==0.2
